* rocDecode/HEVC: Code clean up.  - Moved HEVC standard specific defines out of HevcVideoParser class to a new file.  - Changed some namings to be more specification compliant.  - No functional changes.
